requirements.txt
langchain
langchain_openai
openai
python-dotenv
streamlit
